optimalizacia

+++ setup/symlinks.sh
@@ -1,31 +1,40 @@
 #!/bin/bash
 
+# Načítanie prostredia
 if [ -e ../.env ]; then
     . ../.env
 else
     . ../.env.dev
 fi
 
-# Základná cesta k témam
+# --- THEMES ---
 src_base="../../../themes/"
 dest_base="../www/wordpress/wp-content/themes"
 
-# Načítanie všetkých tém
 for source in ../www/themes/*; do
-    # Skontroluj, či je to priečinok
     [ -d "$source" ] || continue
-
-    # Zisti názov témy
     theme=$(basename "$source")
     dest="$dest_base/$theme"
     source="$src_base$theme"
 
-    # Vymaž existujúce linky/priečinok
-    if [ -L "$dest" ] || [ -d "$dest" ]; then
-        rm -rf "$dest"
-    fi
+    [ -L "$dest" ] || [ -d "$dest" ] && rm -rf "$dest"
 
-    # Vytvor symbolický link
     ln -s "$source" "$dest"
-    echo "Symbolický link vytvorený: $dest -> $source"
+    echo "Symbolický link vytvorený (theme): $dest -> $source"
+done
+
+# --- PLUGINS ---
+src_base="../../../plugins/"
+dest_base="../www/wordpress/wp-content/plugins"
+
+for source in ../www/plugins/*; do
+    [ -d "$source" ] || continue
+    plugin=$(basename "$source")
+    dest="$dest_base/$plugin"
+    source="$src_base$plugin"
+
+    [ -L "$dest" ] || [ -d "$dest" ] && rm -rf "$dest"
+
+    ln -s "$source" "$dest"
+    echo "Symbolický link vytvorený (plugin): $dest -> $source"
 done
